+++ test/integration/cookbooks/xcode-tests/metadata.rb
@@ -1,0 +1,10 @@
+name             'xcode-tests'
+maintainer       'Antek S. Baranski'
+maintainer_email 'antek.baranski@gmail.com'
+license          'Apache 2.0'
+description      'xcode cookbook test harness'
+version          '1.0.0'
+
+supports         'mac_os_x'
+
+depends          'xcode'